:bug: FIX. Breathing 테이블에 측정 값이 안 들어가는 문제 수정

+++ src/main/java/com/hsu/shimpyoo/domain/breathing/entity/Breathing.java
@@ -26,15 +26,15 @@ public class Breathing extends BaseEntity {
     private User userId; // 사용자 기본키
 
     @Column(name = "breathing_rate")
-    private Float breathingRate; // 호흡 수치
+    private Double breathingRate; // 호흡 수치
 
     @Column(name = "first")
-    private Float first; // 1회차
+    private Double first; // 1회차
 
     @Column(name = "second")
-    private Float second; // 2회차
+    private Double second; // 2회차
 
     @Column(name = "third")
-    private Float third; // 3회차
+    private Double third; // 3회차
 }
